backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
passlib[bcrypt]
python-dotenv
httpx
psycopg2-binary
email-validator
fastapi-mail
PyJWT
pydantic-settings
aiofiles
python-multipart
openai
reportlab
pytesseract
PyPDF2
pdf2image
docx
pandas
pdfkit==1.0.0
Markdown==3.10.2
anthropic==0.79.0